added jsp (view): added registering, logging in/out, added possibility showing accounts and credit cards for logged user, added work with account(adding, blocking), added work with credit card(adding, deleting), added command pattern.

+++ src/main/java/com/my/model/dao/GenericDao.java
@@ -5,10 +5,10 @@ import java.util.List;
 public interface GenericDao<T> extends AutoCloseable{
 
     void add(T entity);
-    T findById(Long id);
+    T findById(Integer id);
     List<T> findAll();
     void update(T entity);
-    void delete(Long id);
+    void delete(Integer id);
     void deleteEntity(T entity);
 
     void close();
